% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggerr.R
\docType{package}
\name{ggerr-package}
\alias{ggerr}
\alias{ggerr-package}
\title{ggerr: Easy errorbars to ggplots}
\description{
This package exports stat_err() for dding errorbars to ggplots.
}
\details{
The ggerr package addes errorbars to ggplots.
It defines the new stat: stat_err(), that can use geoms:
errorbar, errorbarh, and point.
See vignette("using_ggerr") for details and examples
}
\author{
\strong{Maintainer}: Thomas Poulsen \email{tapo@novozymes.com}

}
\keyword{internal}
