% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_err.R
\name{stat_err}
\alias{stat_err}
\title{stat_err}
\usage{
stat_err(mapping = NULL, data = NULL, geom = "errorbar",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ..., spread = "sd", mult = 1)
}
\arguments{
\item{mapping}{needs some grouping, eg color or group in aes()}

\item{data}{normally inherited from call to ggplot()}

\item{geom}{use "errorbars" for vertical errorbar, "errorbarh" for horizontal errorbar, "point" for mean}

\item{position}{leave as "identity"}

\item{na.rm}{not used}

\item{show.legend}{std ggplot}

\item{inherit.aes}{std ggplot}

\item{...}{passed on}

\item{spread}{statistics to use: "sd" for standard deviation, "se" for standard error}

\item{mult}{factor to multiply on the "spread". Default 1 means error-bars are +/- 1 sd around the mean}
}
\value{
used in ggplot()
}
\description{
stat_err adds error-bars to scatterplot
}
\details{
Adding a point at the median, and vertical errorbars requirer two call to stat_err(). Adding a horizontal error bar needs one more call.
ggplot(data =  iris, aes(x= Petal.Length, y=Petal.Width, group=Species)) + geom_point(aes(color=Species)) + stat_err(width=.1, spread = "se", mult = 2) + stat_err(geom="point") + stat_err(geom="errorbarh", height=.1, spread = "se", mult = 2)
}
\examples{
ggplot(data =  iris, aes(x= Petal.Length, y=Petal.Width, group=Species)) +
  geom_point(aes(color=Species)) +
  stat_err(width=.1, spread = "se", mult = 2) +
  stat_err(geom="point") +
  stat_err(geom="errorbarh", height=.1, spread = "se", mult = 2)
}
\author{
Thomas Poulsen
}
